+++ tests/testthat/test_AcqFunctionSmsEgo.R
@@ -32,3 +32,30 @@ test_that("AcqFunctionSmsEgo works", {
   expect_named(res)
   expect_setequal(colnames(res), c(acqf$id, "acq_epsilon"))
 })
+
+test_that("AcqFunctionSmsEgo transforms ys_front onto the output trafo scale", {
+  inst = MAKE_INST(OBJ_1D_2, PS_1D, trm("evals", n_evals = 5L))
+  surrogate = SurrogateLearnerCollection$new(
+    list(REGR_FEATURELESS, REGR_FEATURELESS$clone(deep = TRUE)),
+    archive = inst$archive
+  )
+  ot = OutputTrafoStandardize$new()
+  ot$invert_posterior = FALSE
+  surrogate$output_trafo = ot
+  acqf = AcqFunctionSmsEgo$new(surrogate = surrogate)
+
+  design = MAKE_DESIGN(inst)
+  inst$eval_batch(design)
+
+  acqf$surrogate$update()
+  expect_true(surrogate$output_trafo_must_be_considered)
+  acqf$progress = 1
+  acqf$update()
+
+  expected = ot$transform(inst$archive$best()[, inst$archive$cols_y, with = FALSE])
+  for (col in inst$archive$cols_y) {
+    set(expected, j = col, value = expected[[col]] * acqf$surrogate_max_to_min[[col]])
+  }
+  expect_equal(acqf$ys_front, as.matrix(expected))
+  expect_true(all(apply(acqf$ys_front, 2L, max) <= acqf$ref_point))
+})